CU-86dtnj3kj - NEON3 - Overview tab - Implement Card with main assets price data
