+++ src/renderer/src/routes/pages/Portfolio/Overview/ChartCardList.tsx
@@ -10,8 +10,8 @@ type TProps = {
 export const ChartCardList = ({ tokenBalance }: TProps) => {
   const { t } = useTranslation('components', { keyPrefix: 'chartCardList' })
   return (
-    <div className="flex flex-col gap-y-4">
-      <span>{t('title')}</span>
+    <div className="flex flex-col gap-y-4 px-4 items-start">
+      <span className="text-sm text-gray-100">{t('title')}</span>
 
       <div className="flex gap-2 flex-wrap">
         {tokenBalance.map((item, index) => (
