+++ src/renderer/src/routes/pages/Portfolio/Overview/index.tsx
@@ -1,11 +1,10 @@
-import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { TTokenBalance } from '@renderer/@types/query'
 import { BalanceChart } from '@renderer/components/BalanceChart'
 import { Separator } from '@renderer/components/Separator'
 import { NumberHelper } from '@renderer/helpers/NumberHelper'
 import { useAccountsSelector } from '@renderer/hooks/useAccountSelector'
 import { useBalances } from '@renderer/hooks/useBalances'
+import { useFilteredBalance } from '@renderer/hooks/useFilteredBalance'
 import { useCurrencySelector } from '@renderer/hooks/useSettingsSelector'
 import { useWalletsSelector } from '@renderer/hooks/useWalletSelector'
 
@@ -18,25 +17,7 @@ export const PortfolioOverviewPage = () => {
   const { currency } = useCurrencySelector()
   const balances = useBalances(accounts)
 
-  const filteredTokenBalances = useMemo(() => {
-    const tokensBalances = balances.data.map(balance => balance.tokensBalances).flat()
-
-    return tokensBalances
-      .filter(balance => balance.exchangeAmount > 0)
-      .reduce((acc, balance) => {
-        const repeated = acc.find(item => item.token.hash === balance.token.hash)
-        if (repeated) {
-          repeated.amountNumber += balance.amountNumber
-          repeated.exchangeAmount += balance.exchangeAmount
-          repeated.amount = NumberHelper.removeLeadingZero(repeated.amountNumber.toFixed(repeated.token.decimals))
-          return acc
-        }
-
-        acc.push(balance)
-        return acc
-      }, [] as TTokenBalance[])
-      .sort((token1, token2) => token2.exchangeAmount - token1.exchangeAmount)
-  }, [balances.data])
+  const filteredTokenBalances = useFilteredBalance(balances)
 
   return (
     <section className="w-full flex flex-col bg-gray-800 rounded shadow-lg py-3 h-full px-4 min-w-0">
@@ -61,12 +42,12 @@ export const PortfolioOverviewPage = () => {
         </div>
       </div>
 
-      <ul className="flex h-[50%] items-center px-20">
+      <ul className="flex h-[50%] items-center px-12">
         <BalanceChart balances={balances} tokenBalance={filteredTokenBalances} />
       </ul>
 
       {filteredTokenBalances.length > 0 && (
-        <div className="flex flex-col gap-y-4">
+        <div className="flex flex-col gap-y-4 items-center">
           <Separator />
           <ChartCardList tokenBalance={filteredTokenBalances.slice(0, 4)} />
         </div>
